Move GLSL into its own module.
